requirements.txt
kafka-python==2.0.1
py4j==0.10.7
pyspark==2.4.5